Wrappers

+++ finalProject/src/main/java/htp/dao/ProductRepository.java
@@ -5,7 +5,7 @@ import htp.entities.db_entities.Product;
 import java.util.List;
 
 public interface ProductRepository extends GenericDao<Product, Long> {
-     Product findByProductCode(Long productCode);
+     Product findByProductCode(Integer productCode);
 
      List<Product> findAll(int limit, int offset);
 }
